+++ packages/builder/lib/src/data/datasources/local_component_service.dart
@@ -76,7 +76,7 @@ class LocalComponentService {
   Future<GetPageResponseEntity> getPreloadedComponent(
       String componentName) async {
     final res = _preloadFile.customJson ??
-        jsonDecode(await rootBundle.loadString('assets/theta_preload.json'));
+        jsonDecode(await rootBundle.loadString(thetaPreloadFilePath));
     return GetPageResponseEntity.fromJson(
         jsonDecode(decompressAndDecrypt(_clientToken.key, res[componentName])));
   }