+++ packages/cli/lib/src/data/repositories/directory_repository_impl.dart
@@ -21,4 +21,14 @@ class DirectoryRepositoryImpl implements DirectoryRepository {
       return Left(Exception(e.toString()));
     }
   }
+
+  @override
+  Future<Either<Exception, void>> preloadImages(
+      Map<String, dynamic> json) async {
+    try {
+      return Right(await _directoryService.preloadImages(json));
+    } catch (e) {
+      return Left(Exception(e.toString()));
+    }
+  }
 }