+++ packages/widgets/lib/src/elements/widgets/image.dart
@@ -5,6 +5,8 @@
 
 // Package imports:
 // Flutter imports:
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:theta_models/theta_models.dart';
@@ -39,9 +41,13 @@ class OpenWImage extends StatelessWidget {
       context: context,
       loop: nodeState.loop,
     ) as String;
-    final result = img.isNotEmpty
+    var result = img.isNotEmpty
         ? img
         : 'https://fftefqqvfkkewuokofds.supabase.co/storage/v1/object/public/theta-assets/cover-min.png';
+    if (state.isPreloaded) {
+      result =
+          'assets/theta_assets/${base64.encode(utf8.encode(nodeState.node.id))}.${result.split('.').last}';
+    }
     final data =
         const NodeOverrideExecuter().executeImage(context, nodeState, result);
     return Container(
@@ -70,10 +76,15 @@ class OpenWImage extends StatelessWidget {
         context: context,
         isWidth: false,
       ),
-      child: Image.network(
-        data,
-        fit: boxFit.value,
-      ),
+      child: state.isPreloaded
+          ? Image.asset(
+              data,
+              fit: boxFit.value,
+            )
+          : Image.network(
+              data,
+              fit: boxFit.value,
+            ),
     );
   }
 }